Renamed the temp to the original one
