+++ learning_projects/MovieAppCleanArchitecture/app/src/main/java/com/example/movieappcleanarchitecture/data/remote/model/MovieApiModel.kt
@@ -1,4 +1,4 @@
-package com.example.movieappcleanarchitecture.datatemp.remote.model
+package com.example.movieappcleanarchitecture.data.remote.model
 
 import com.google.gson.annotations.SerializedName
