+++ learning_projects/MovieAppCleanArchitecture/app/src/main/java/com/example/movieappcleanarchitecture/presentation/components/MovieGrid.kt
@@ -8,7 +8,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.movieappcleanarchitecture.datatemp.model.Movie
+import com.example.movieappcleanarchitecture.data.model.Movie
 
 @Composable
 fun MovieGrid(movieList: List<Movie>) {
